// packages/test/generatedCode/identity/userFlowAttributes/item/identityUserFlowAttributeItemRequestBuilderGetRequestConfiguration.ts
import {IdentityUserFlowAttributeItemRequestBuilderGetQueryParameters} from './identityUserFlowAttributeItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface IdentityUserFlowAttributeItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: IdentityUserFlowAttributeItemRequestBuilderGetQueryParameters;
}